+++ src/App.jsx
@@ -3,15 +3,15 @@ import { AnimatePresence } from 'framer-motion';
 import Navigation from './components/layout/Navigation';
 import { ThemeProvider } from './context/ThemeContext';
 import { useTheme } from './context/ThemeContext';
-import { BlogProvider } from './components/BlogProvider';
+import { BlogProvider } from './components/blog/BlogProvider.jsx';
 import Home from './pages/Home';
 import Projects from './pages/Projects';
 import Contact from './pages/Contact';
 import WordWize from './components/games/WordWize';
 import WizeSnake from './components/games/WizeSnake';
-import PageTransition from './components/PageTransition';
+import PageTransition from './components/layout/PageTransition.jsx';
 import Newsletter from './pages/Newsletter';
-import ThemeIcon from './components/ThemeIcon';
+import ThemeIcon from './components/common/ThemeIcon.jsx';
 import LearningAssistant from './pages/LearningAssistant';
 import PrivacyPolicy from './pages/PrivacyPolicy';
 import TermsOfService from './pages/TermsOfService';
