+++ src/components/blog/BlogProvider.jsx
@@ -1,5 +1,5 @@
 import { createContext, useContext, useState } from 'react';
-import blogPosts from '../data/blogPosts';
+import blogPosts from '../../data/blogPosts.jsx';
 
 const BlogContext = createContext();
