+++ src/components/embeds/SubstackFeed.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import { motion } from 'framer-motion';
-import { useTheme } from '../context/ThemeContext';
+import { useTheme } from '../../context/ThemeContext.jsx';
 
 export default function SubstackFeed() {
   const { theme } = useTheme();
