+++ src/components/forms/ContactForm.jsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { useTheme } from '../context/ThemeContext';
+import { useTheme } from '../../context/ThemeContext.jsx';
 import { motion } from 'framer-motion';
 import emailjs from '@emailjs/browser';
