+++ src/components/marketing/FeaturedProject.jsx
@@ -1,5 +1,5 @@
 import { motion } from 'framer-motion';
-import { useTheme } from '../context/ThemeContext';
+import { useTheme } from '../../context/ThemeContext.jsx';
 
 function FeaturedProject({ project, index }) {
   const { theme } = useTheme();
