+++ src/components/marketing/MissionCarousel.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import { motion, AnimatePresence } from 'framer-motion';
-import { useTheme } from '../context/ThemeContext';
+import { useTheme } from '../../context/ThemeContext.jsx';
 
 function MissionCarousel() {
   const { theme } = useTheme();
