+++ src/pages/AboutUs.jsx
@@ -1,5 +1,5 @@
 import { useTheme } from '../context/ThemeContext';
-import PageTransition from '../components/PageTransition';
+import PageTransition from '../components/layout/PageTransition.jsx';
 import { motion } from 'framer-motion';
 import { Link } from 'react-router-dom';
